Add gallery files via upload
A sample gallery of images before trying to update with a modal gallery and a possible text overlay. At this point, I'm having issues with making the images all the same size and centering them, as well as some questions regarding making these images responsive (but this is probably the last step in my slew of problems!) Would love to first use css floats for this, rather than flexbox.